Extract defineDefaultParameterTypes
