--- javascript/src/Expression.ts
@@ -1,6 +1,0 @@
-import Argument from './Argument.js'
-
-export default interface Expression {
-  readonly source: string
-  match(text: string): readonly Argument[] | null
-}
