+++ javascript/src/index.ts
@@ -1,13 +1,13 @@
 import Argument from './Argument.js'
 import CucumberExpression from './CucumberExpression.js'
 import CucumberExpressionGenerator from './CucumberExpressionGenerator.js'
-import Expression from './Expression.js'
 import ExpressionFactory from './ExpressionFactory.js'
 import GeneratedExpression from './GeneratedExpression.js'
 import Group from './Group.js'
 import ParameterType from './ParameterType.js'
 import ParameterTypeRegistry from './ParameterTypeRegistry.js'
 import RegularExpression from './RegularExpression.js'
+import { Expression } from './types.js'
 
 export {
   Argument,
