+++ javascript/src/types.ts
@@ -1,0 +1,11 @@
+import Argument from './Argument.js'
+import ParameterType from './ParameterType'
+
+export interface DefinesParameterType {
+  defineParameterType<T>(parameterType: ParameterType<T>): void
+}
+
+export interface Expression {
+  readonly source: string
+  match(text: string): readonly Argument[] | null
+}
